New verson without simulating key strokes
This new version uses the switcher programmatically, but to achieve this I had to copy all the files involve. It's not necessary the entire files, but anyway. The files are from 3.36 gnome-shell version. And I didn't change the indentation, so there's a mixture between tabs and spaces, but I don't care.